+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/TransferManager.cs
@@ -10,6 +10,7 @@ using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using Azure.Core;
+using Azure.Core.Pipeline;
 using Azure.Storage.DataMovement.Models;
 using Azure.Storage.DataMovement.Models.JobPlan;
 
@@ -89,6 +90,8 @@ namespace Azure.Storage.DataMovement
         internal ArrayPool<byte> UploadArrayPool => _arrayPool;
         private ArrayPool<byte> _arrayPool;
 
+        internal ClientDiagnostics ClientDiagnostics { get; }
+
         /// <summary>
         /// Protected constructor for mocking.
         /// </summary>
@@ -128,6 +131,7 @@ namespace Azure.Storage.DataMovement
             _dataTransfers = new Dictionary<string, DataTransfer>();
             _arrayPool = ArrayPool<byte>.Shared;
             _errorHandling = options?.ErrorHandling != default ? options.ErrorHandling : ErrorHandlingBehavior.StopOnAllFailures;
+            ClientDiagnostics = new ClientDiagnostics(options?.ClientOptions ?? ClientOptions.Default);
         }
 
         #region Job Channel Management
@@ -361,6 +365,11 @@ namespace Azure.Storage.DataMovement
             CancellationToken cancellationToken = default)
         {
             CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
+            Argument.AssertNotNullOrWhiteSpace(transferId, nameof(transferId));
+            Argument.AssertNotNull(sourceResource, nameof(sourceResource));
+            Argument.AssertNotNull(destinationResource, nameof(destinationResource));
+
+            transferOptions ??= new TransferOptions();
 
             if (_dataTransfers.ContainsKey(transferId))
             {
@@ -429,33 +438,15 @@ namespace Azure.Storage.DataMovement
 
             transferOptions ??= new TransferOptions();
 
-            string transferId;
-            bool resumeJob = false;
-            if (!string.IsNullOrEmpty(transferOptions.ResumeFromCheckpointId))
-            {
-                resumeJob = true;
-                transferId = transferOptions.ResumeFromCheckpointId;
-
-                if (_dataTransfers.ContainsKey(transferId))
-                {
-                    // Remove the stale DataTransfer so we can pass a new DataTransfer object
-                    // to the user and also track the transfer from the DataTransfer object
-                    _dataTransfers.Remove(transferId);
-                }
-            }
-            else
-            {
-                // Add Transfer to Checkpointer
-                transferId = Guid.NewGuid().ToString();
-                await _checkpointer.AddNewJobAsync(transferId, _cancellationToken).ConfigureAwait(false);
-            }
+            string transferId = Guid.NewGuid().ToString();
+            await _checkpointer.AddNewJobAsync(transferId, _cancellationToken).ConfigureAwait(false);
 
             DataTransfer dataTransfer = await BuildAndAddTransferJobAsync(
                 sourceResource,
                 destinationResource,
                 transferOptions,
                 transferId,
-                resumeJob,
+                false,
                 cancellationToken).ConfigureAwait(false);
 
             return dataTransfer;
@@ -529,7 +520,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -568,7 +560,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         CheckPointFolderPath: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -600,7 +593,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -645,7 +639,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -691,7 +686,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {
@@ -730,7 +726,8 @@ namespace Azure.Storage.DataMovement
                         queueChunkTask: QueueJobChunkAsync,
                         checkpointer: _checkpointer,
                         errorHandling: _errorHandling,
-                        arrayPool: _arrayPool);
+                        arrayPool: _arrayPool,
+                        clientDiagnostics: ClientDiagnostics);
 
                     if (resumeJob)
                     {