+++ sdk/storage/Azure.Storage.DataMovement/tests/DownloadChunkHandlerTests.cs
@@ -11,6 +11,7 @@ using System.Threading;
 using Azure.Core;
 using Azure.Storage.DataMovement.Models;
 using Azure.Storage.Tests.Shared;
+using Azure.Core.Pipeline;
 
 namespace Azure.Storage.DataMovement.Tests
 {
@@ -25,6 +26,8 @@ namespace Azure.Storage.DataMovement.Tests
         private readonly string _copyToChunkFileMsg = "Amount of Copy To Chunk File Task calls were incorrect.";
         private readonly string _reportProgressInBytesMsg = "Amount of Progress amount calls were incorrect.";
         private readonly string _completeFileDownloadMsg = "Complete File Download call amount calls were incorrect.";
+
+        private ClientDiagnostics ClientDiagnostics => new(ClientOptions.Default);
 
         private void VerifyDelegateInvocations(
             MockDownloadChunkBehaviors behaviors,
@@ -217,6 +220,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -263,6 +267,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -330,6 +335,7 @@ namespace Azure.Storage.DataMovement.Tests
                     QueueCompleteFileDownload = mockBehaviors.QueueCompleteFileDownloadTask.Object,
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -388,6 +394,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -456,6 +463,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             List<Task> runningTasks = new List<Task>();
@@ -513,6 +521,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -560,6 +569,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -606,6 +616,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             PredictableStream content = new PredictableStream(blockSize);
@@ -652,6 +663,7 @@ namespace Azure.Storage.DataMovement.Tests
                     ReportProgressInBytes = mockBehaviors.ReportProgressInBytesTask.Object,
                     InvokeFailedHandler = mockBehaviors.InvokeFailedEventHandlerTask.Object,
                 },
+                ClientDiagnostics,
                 cancellationToken: CancellationToken.None);
 
             // Act